refactor: db 맞추기

+++ src/main/java/com/factory/chatbot/infrastructure/repository/AnomalyLogRepository.java
@@ -8,10 +8,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AnomalyLogRepository extends JpaRepository<AnomalyLog, Long> {
 
-    /**
-     * 설비 ID로 조회하여 발생 시각(OccurredTime) 기준 내림차순으로 최근 5개만 가져오는 매직 메서드
-     */
-    List<AnomalyLog> findTop5ByEquipmentIdOrderByOccurredTimeDesc(Integer equipmentId);
+    List<AnomalyLog> findTop5ByEquipmentIdOrderByOccurredTimeDesc(Long equipmentId);
 
-    List<AnomalyLog> findByEquipmentIdOrderByOccurredTimeDesc(Integer equipmentId);
+    List<AnomalyLog> findByEquipmentIdOrderByOccurredTimeDesc(Long equipmentId);
 }
